requirements.txt
spotipy==2.22.0
python-dotenv==1.0.0